backend/requirements.txt
fastapi>=0.115.0
python-multipart>=0.0.9
uvicorn[standard]>=0.32.0
python-dotenv>=1.0.0
pydantic>=2.0.0
psycopg2-binary>=2.9.9
google-generativeai>=0.8.0
openpyxl>=3.1.0
pytest>=8.0.0
